LayoutTests/fast/css/inherited-properties-rare-text.html

<script>
if (window.testRunner)
    testRunner.dumpAsText();
</script>
<style>
.a { -webkit-font-feature-settings: 'dlig'; -webkit-font-smoothing: antialiased; -webkit-text-orientation: upright; }
</style>
<div class=a>
    <div id=test1></div>
</div>
<div id=test2></div>
<script>
function test(e, p) {
    var testDiv = document.getElementById(e);
    var cssValue = window.getComputedStyle(testDiv).getPropertyCSSValue(p);
    document.write(e + " " + p + ": " + cssValue.cssText + "<br>");
}
    
test('test1', '-webkit-font-feature-settings');
test('test2', '-webkit-font-feature-settings');
test('test1', '-webkit-font-smoothing');
test('test2', '-webkit-font-smoothing');
test('test1', '-webkit-text-orientation');
test('test2', '-webkit-text-orientation');
</script>
