LayoutTests/fast/css/inherited-properties-rare-text-expected.txt

test1 font-feature-settings: "dlig"
test2 font-feature-settings: normal
test1 -webkit-font-smoothing: antialiased
test2 -webkit-font-smoothing: auto
test1 -webkit-text-orientation: upright
test2 -webkit-text-orientation: mixed
